w06-1 Part-1 Refactor Input and Label into a Re-usable Component
